requirements.txt
anyio==3.7.0
catboost==1.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
Django==3.2
django-cors-headers==4.1.0
djangorestframework==3.12.4
drf-yasg==1.21.5
exceptiongroup==1.1.1
fastapi==0.97.0
flake8==6.0.0
fonttools==4.40.0
graphviz==0.20.1
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-resources==5.12.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
mccabe==0.7.0
numpy==1.24.3
packaging==23.1
pandas==1.5.3
Pillow==9.5.0
plotly==5.15.0
pycodestyle==2.10.0
pydantic==1.10.9
pyflakes==3.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
tenacity==8.2.2
threadpoolctl==3.1.0
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.3
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
whitenoise==6.5.0
zipp==3.15.0
